refactor download functions into new file
