+++ src/components/DicomData/TableComponents/DicomTable.tsx
@@ -1,6 +1,8 @@
 import React, { useState } from "react";
 import logger from "../../utils/Logger";
-import { tagUpdater, downloadDicomFile, createFile } from "../TagUpdater";
+import { tagUpdater } from "../TagUpdater";
+import { createFile } from "../DownloadFuncs";
+import { downloadDicomFile } from "../DownloadFuncs";
 import { DicomTableProps } from "../../../types/DicomTypes";
 import { useFilteredRows } from "../../Hooks/useFilteredRows";
 import { createRows } from "./rowUtils";
